+++ src/Actions/GetSubAction.php
@@ -2,7 +2,6 @@
 
 namespace Yomafleet\CognitoAuthenticator\Actions;
 
-use Illuminate\Http\Request;
 use Yomafleet\CognitoAuthenticator\JwtDecoder;
 use Illuminate\Validation\UnauthorizedException;
 use Yomafleet\CognitoAuthenticator\Models\UserPool;
@@ -12,12 +11,11 @@ use Yomafleet\CognitoAuthenticator\Factories\UserPoolFactory;
 use Yomafleet\CognitoAuthenticator\Contracts\CanGetSubContract;
 use Yomafleet\CognitoAuthenticator\Contracts\TokenFactoryContract;
 use Yomafleet\CognitoAuthenticator\Contracts\UserPoolFactoryContract;
-use Yomafleet\CognitoAuthenticator\Exceptions\AuthorizationHeaderNotFoudException;
 
 class GetSubAction implements CanGetSubContract
 {
-    /** @var \Illuminate\Http\Request */
-    protected $request;
+    /** @var string */
+    protected $authToken;
 
     /** @var \Yomafleet\CognitoAuthenticator\Contracts\UserPoolFactoryContract */
     protected $userPoolFactory;
@@ -26,11 +24,11 @@ class GetSubAction implements CanGetSubContract
     protected $tokenFactory;
 
     public function __construct(
-        Request $request,
+        string $authToken,
         UserPoolFactoryContract $userPoolFactory = null,
         TokenFactoryContract $tokenFactory = null,
     ) {
-        $this->request = $request;
+        $this->authToken = $authToken;
         $this->userPoolFactory = $userPoolFactory ?: new UserPoolFactory();
         $this->tokenFactory = $tokenFactory ?: new TokenFactory();
     }
@@ -56,30 +54,13 @@ class GetSubAction implements CanGetSubContract
     {
         $pool = $this->getUserPool(['']);
         $decoder = $this->getDecoder($pool);
-        $token = $decoder->decode($this->getBearer());
+        $token = $decoder->decode($this->authToken);
 
         if (! $token) {
             throw new UnauthorizedException();
         }
 
         return $token->getSub();
-    }
-
-    /**
-     * Get Bearer token for authroization header
-     *
-     * @throws \Yomafleet\CognitoAuthenticator\Exceptions\AuthorizationHeaderNotFoudException
-     * @return string
-     */
-    protected function getBearer(): string
-    {
-        $authorization = $this->request->header('authorization');
-
-        if (! $authorization) {
-            throw new AuthorizationHeaderNotFoudException();
-        }
-
-        return trim(str_ireplace('Bearer ', '', $authorization));
     }
 
     /**
